Update `.emplace()` methods to match spec
- It is possible that `map.get(key)` could have a value of `undefined`, so do need to check `map.has(key)` instead.
- Handle case where `map.has(key)` but no update function given

I'm returning `map.get(key)` for cases where `map.has(key)` but no update callback is given. No instructions are given for this case, but the document gives reason why it should never return `undefined`... So it should return *something*, and the current value is the only thing that makes any sense.